third_party/WebKit/LayoutTests/fast/table/table-rowspan-row-height-less-than-content-height.html

<!DOCTYPE html>
<style>
    table { width : 600px }
    #percent-height { height : 1% }
    td { font: 15px/1 Ahem }
</style>
<script src="../../resources/check-layout.js"></script>
<body onload="checkLayout('tr');">
    <h3>Test for chromium bug : <a href="https://code.google.com/p/chromium/issues/detail?id=406043">406043</a>. Negative row height when cell has percentage height.</h3>
    <table>
        <tr data-offset-y="2">
            <td height="100px" rowspan="4"> This is rowspan cell. It would span 3 rows in table.</td>
            <td width="500px"> This is first row second column. </td>
        </tr>
        <tr data-offset-y="21">
            <td id="percent-height"> This is second row first column. But it should display as second column. </td>
        </tr>
        <tr data-offset-y="70">
            <td id="percent-height"> This is third row first column. But it should display as second column. </td>
        </tr>
        <tr data-offset-y="119">
            <td id="percent-height"> This is fourth row first column. But it should display as second column. </td>
        </tr>
    </table>
</body>
